requirements.txt
Django==4.1
numpy==1.23.3
Pillow==9.2.0
requests==2.28.1
django-bootstrap4==22.2
python-multipart